% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ff_history.R
\name{summary.ff_history}
\alias{summary.ff_history}
\title{Compute Pretty Summary Table}
\usage{
\method{summary}{ff_history}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments passed on to the methods (currently not used).}
}
\description{
Uses the R package gt to create a pretty html table of the
ff_history summary data frame.
}
\examples{
\donttest{
conn <- ffscrapr::sleeper_connect(2023, "957091261993811968")
ff_history_obj <- ff_history(conn)

conn <- ffscrapr::mfl_connect(2022, 54040)

# Create Summary Tables
tbl <- summary(ff_history_obj)

}
}
